Hid previous page contents, began work on first page of Wizard
The previous contents of the page, which appeared like a character sheet
have been hidden using CSS.  In their place I have begun work on the
Wizard system that is to partially replace that.  Currently the first
page of the UI is built, though it does nothing as of yet.