Remaining getter methods for packageFile completed. dataFile class defined.
packageFile now has getter methods for Licenses, Maintainers, and
Resources.  It may be possible to make a method that combines License
and Licenses into one pre-formatted output.  It may be worth making a
method that gives a pre-formatted output for Maintainers as well.

The first version of the dataFiles class was also defined in this
commit.  Currently it only keeps track of the data it was passed. The
datafiles will be lazy initialized the first time their data is
requested.