Removing test from load.json
The last line of load.json tried to load a package named test.  This
package does not exist and was only there to test the error handling
code of the package loader.  Since it appears to work, I am removing
this line.  There is no reason to do an extra GET request anymore.