Fixed minor bug in path for dataFile class and added getData method.
There was a minor bug in the path calculation for the dataFile class
that resulted in the .json file extension being applied twice, which has
now been resolved.  Additionally the getData method on the dataFile
class has been implemented.  It returns a promise object that can be
acted upon once the data has loaded.