First advanced option added, support for advanced options improved
Toggling on and off advanced options now has a noticible effect as it
hides any sections of the page that have been marked as advanced options.
The first advanced option has been added in the form of "Generate Last
Name", which toggles whether or not the name generator includes a last
name.