Put together a basic draft of ability scores interface
The first version of the ability scores interface has been created.
It has room for ability scores to be entered manually, an area where rolled
scores will appear and options for selecting your preferred rolling method.
The current interface is non-functional though and exists primarily as a
draft of what the interface might look like.  Some discussions will need
to be had to determine the actual layout that will be used.